Move server into central CLI command
